requirements.txt
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=0.19.0
rich>=13.0.0
inquirer>=3.1.0
requests>=2.31.0
Pillow>=10.0.0

# ChromaDB dependencies
langchain-openai>=0.0.5
langchain-chroma>=0.0.5
chromadb>=0.4.22

# Optional document handling dependencies
python-docx>=0.8.11
PyPDF2>=3.0.0
odfpy>=1.4.1
pyth>=0.7.0

# Note: Some dependencies might need to be installed manually if not available on PyPI
# pyth might need to be installed from source as it's not actively maintained